Namespace and DatabricksQueryOptions
