+++ source/ProcessManager.Orchestrations/Processes/BRS_021/ElectricalHeatingCalculation/V1/Orchestration/Steps/EnqueueActorMessagesStep.cs
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Core.Databricks.SqlStatementExecution;
-using Energinet.DataHub.ProcessManager.Components.Databricks.SqlStatementApi;
 using Energinet.DataHub.ProcessManager.Core.Domain.OrchestrationInstance;
 using Energinet.DataHub.ProcessManager.Orchestrations.Processes.BRS_021.ElectricalHeatingCalculation.V1.Model;
 using Energinet.DataHub.ProcessManager.Shared.Processes.Activities;
